feat(cache): implement dynamic cache expiration in Redis
Introduce a new aspect `CacheTtlAspect` to dynamically set the time-to-live (TTL)for Redis cache entries based on the `CacheTtl` annotation. This allows for more
flexible cache management, where the TTL can be specified per method or class.

The `CacheTtlContext` class has been introduced to manage the TTL values in a
thread-safe manner using a thread-local stack. This enables the support for
nested cache operations with different TTLs.

The `WeiRedisCache` and `WeiRedisCacheManager` classes have been updated toutilize the TTL from the `CacheTtlContext` when setting cache entries, ensuring
that the dynamic TTL values are applied correctly.

BREAKING CHANGE: The cache expiration behavior has been changed to support
dynamic TTLs. If your application relies on the previous static TTL behavior,
you will need to adjust your configurations or annotations to accommodate this
change.